fix(api-server): handle path names with 'graphql' (#157)
We have code in the api-server package that looks for all serverless
functions and automatically loads them and adds them to the api server.
We want the first function in the list to be the graphql function. But
the code we had to find the graphql function looked at the entire path
for the 'graphql' string. So if the project was placed in a parent
folder that had 'graphql' in its name *all* functions would match that
check.

The updated code in this PR only looks at the file name, not the entire
path.

Fix ported over and improved from
https://github.com/redwoodjs/graphql/pull/12055