fix(supertokens): Fix whitespace handling in setup script (#158)
I was upgrading to Vitest 3.x and noticed one of our snapshot tests
started failing. Looking into it I realized we weren't handling
whitespace properly when setting up Supertokens auth. This PR fixes
that.

I'm guessing this is the Vitest change that caused the diff in
snapshots: https://github.com/vitest-dev/vitest/pull/7400
It doesn't say in the PR description, but the [release notes for Vitest
3.1.0](https://github.com/vitest-dev/vitest/releases/tag/v3.1.0), where
it was first included, says
> This change can cause small amount of very old snapshots to be
updated, but there will be no functional change to how they work.