feat(esm): `@cedarjs/testing` dual ESM/CJS build (#163)
The testing package is a bit weird in that it has both a `src/`
directory and a `config/` directory with code. Plus `api/`, `cache/` and
`/web` with just an index file and a package.json. I know what to do
with the `src/` directory - same as I've done with all the other ESM/CJS
conversions. But not sure what to do with the rest of the directories
yet.

⚠️ The goal is that this change should be non-breakable, but if you're
doing dist imports from `@cedarjs/testing` it might be. Please let me
know in that case and I'll fix it.