chore(tests): Await rejects.toThrowErrorMatchingInlineSnapshot (#168)
I was seeing this error on CI
https://cloud.nx.app/runs/eLSbQ5d7FY/task/%40cedarjs%2Fvite%3Atest

```
Promise returned by `expect(actual).rejects.toThrowErrorMatchingInlineSnapshot(expected)` was not awaited. Vitest currently auto-awaits hanging assertions at the end of the test, but this will cause the test to fail in Vitest 3. Please remember to await the assertion.
    at /home/runner/work/cedar/cedar/packages/vite/src/plugins/__tests__/vite-plugin-rsc-route-auto-loader.test.ts:332:21
```

This PR properly adds `await`