chore(jobs): Move logic for runAt calculation to Executor (#179)
It should be the responsibility of the Executor to calculate when the
job should be next executed in case of an error.
I prefer the adaptors be more "dumb". It makes them easier to implement
across different storage solutions. And it makes the behavior more
consistent across the board too.