feat(testing): Surface user errors inside MockProviders (#188)
Surface more errors to the end user to help them debug issues during
testing.

The MockProviders pull in a lot of code from the user's project, and if
there are any errors in that code we'd swallow it and it'd just surface
as `TypeError: _router.routes.project is not a function` or similar.
That's not very helpful.

Now I print the error with `console.warn` instead, so the user can see
it