fix(api-server): lambdaLoader: improve handler import logic (#190)
For ESModule and CommonJS compatibility.
When using a bundler like Rollup, the handler function is exported as
`default`.

This change defaults to a named `handler` export, just like before. But
if an export like that can't be found it'll now also look for a
`default` export with a `handler` property on it.

All credit for this one goes to @richard-stafflink who created the
original PR over at https://github.com/redwoodjs/graphql/pull/12064

---------

Co-authored-by: Richard J <106497642+richard-stafflink@users.noreply.github.com>